Fixes (#2)
* Fixes

* Uncomment TestMain

* Lint
