+++ server/envs/regtest.env
@@ -7,3 +7,4 @@ ARK_DATADIR=./data/regtest
 ARK_ESPLORA_URL=http://localhost:3000
 ARK_WALLET_ADDR=localhost:6060
 ARK_DB_URL=postgresql://root:secret@127.0.0.1:5432/ark-db?sslmode=disable
+ARK_LIVE_STORE_TYPE=inmemory
