+++ server/internal/core/application/utils.go
@@ -3,12 +3,9 @@ package application
 import (
 	"bytes"
 	"context"
-	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
-	"sort"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/ark-network/ark/common"
@@ -21,448 +18,6 @@ import (
 	"github.com/decred/dcrd/dcrec/secp256k1/v4"
 	log "github.com/sirupsen/logrus"
 )
-
-type timedTxRequest struct {
-	domain.TxRequest
-	boardingInputs []ports.BoardingInput
-	timestamp      time.Time
-	musig2Data     *tree.Musig2
-}
-
-func (t timedTxRequest) hashID() [32]byte {
-	return sha256.Sum256([]byte(t.Id))
-}
-
-type txRequestsQueue struct {
-	lock          *sync.RWMutex
-	requests      map[string]*timedTxRequest
-	vtxos         map[string]struct{}
-	vtxosToRemove []string
-}
-
-func newTxRequestsQueue() *txRequestsQueue {
-	lock := &sync.RWMutex{}
-	requestsById := make(map[string]*timedTxRequest)
-	vtxos := make(map[string]struct{})
-	vtxosToRemove := make([]string, 0)
-	return &txRequestsQueue{lock, requestsById, vtxos, vtxosToRemove}
-}
-
-func (m *txRequestsQueue) len() int64 {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-
-	count := int64(0)
-	for _, p := range m.requests {
-		if len(p.Receivers) > 0 {
-			count++
-		}
-	}
-	return count
-}
-
-func (m *txRequestsQueue) push(
-	request domain.TxRequest,
-	boardingInputs []ports.BoardingInput,
-	musig2Data *tree.Musig2,
-) error {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	if _, ok := m.requests[request.Id]; ok {
-		return fmt.Errorf("duplicated tx request %s", request.Id)
-	}
-
-	for _, input := range request.Inputs {
-		for _, pay := range m.requests {
-			for _, pInput := range pay.Inputs {
-				if input.Txid == pInput.Txid && input.VOut == pInput.VOut {
-					return fmt.Errorf("duplicated input, %s:%d already used by tx request %s", input.Txid, input.VOut, pay.Id)
-				}
-			}
-		}
-	}
-
-	for _, input := range boardingInputs {
-		for _, request := range m.requests {
-			for _, pBoardingInput := range request.boardingInputs {
-				if input.Txid == pBoardingInput.Txid && input.VOut == pBoardingInput.VOut {
-					return fmt.Errorf("duplicated boarding input, %s:%d already used by tx request %s", input.Txid, input.VOut, request.Id)
-				}
-			}
-		}
-	}
-
-	now := time.Now()
-	m.requests[request.Id] = &timedTxRequest{request, boardingInputs, now, musig2Data}
-	for _, vtxo := range request.Inputs {
-		if vtxo.IsNote() {
-			continue
-		}
-		m.vtxos[vtxo.String()] = struct{}{}
-	}
-	return nil
-}
-
-func (m *txRequestsQueue) pop(num int64) []timedTxRequest {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	requestsByTime := make([]timedTxRequest, 0, len(m.requests))
-	for _, p := range m.requests {
-		// Skip tx requests without registered receivers.
-		if len(p.Receivers) <= 0 {
-			continue
-		}
-
-		requestsByTime = append(requestsByTime, *p)
-	}
-	sort.SliceStable(requestsByTime, func(i, j int) bool {
-		return requestsByTime[i].timestamp.Before(requestsByTime[j].timestamp)
-	})
-
-	if num < 0 || num > int64(len(requestsByTime)) {
-		num = int64(len(requestsByTime))
-	}
-
-	result := make([]timedTxRequest, 0, num)
-
-	for _, p := range requestsByTime[:num] {
-		result = append(result, p)
-		for _, vtxo := range m.requests[p.Id].Inputs {
-			m.vtxosToRemove = append(m.vtxosToRemove, vtxo.String())
-		}
-		delete(m.requests, p.Id)
-	}
-
-	return result
-}
-
-func (m *txRequestsQueue) update(request domain.TxRequest, musig2Data *tree.Musig2) error {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	r, ok := m.requests[request.Id]
-	if !ok {
-		return errTxRequestNotFound{request.Id}
-	}
-
-	// sum inputs = vtxos + boarding utxos + notes + recovered vtxos
-	sumOfInputs := uint64(0)
-	for _, input := range request.Inputs {
-		sumOfInputs += input.Amount
-	}
-
-	for _, boardingInput := range r.boardingInputs {
-		sumOfInputs += boardingInput.Amount
-	}
-
-	// sum outputs = receivers VTXOs
-	sumOfOutputs := uint64(0)
-	for _, receiver := range request.Receivers {
-		sumOfOutputs += receiver.Amount
-	}
-
-	if sumOfInputs != sumOfOutputs {
-		return fmt.Errorf("sum of inputs %d does not match sum of outputs %d", sumOfInputs, sumOfOutputs)
-	}
-
-	r.TxRequest = request
-
-	if musig2Data != nil {
-		r.musig2Data = musig2Data
-	}
-	return nil
-}
-
-func (m *txRequestsQueue) delete(ids []string) error {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	for _, id := range ids {
-		req, ok := m.requests[id]
-		if !ok {
-			continue
-		}
-		for _, vtxo := range req.Inputs {
-			delete(m.vtxos, vtxo.String())
-		}
-		delete(m.requests, id)
-	}
-	return nil
-}
-
-func (m *txRequestsQueue) deleteAll() error {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	m.requests = make(map[string]*timedTxRequest)
-	m.vtxos = make(map[string]struct{})
-	return nil
-}
-
-func (m *txRequestsQueue) deleteVtxos() {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-	for _, vtxo := range m.vtxosToRemove {
-		delete(m.vtxos, vtxo)
-	}
-	m.vtxosToRemove = make([]string, 0)
-}
-
-func (m *txRequestsQueue) viewAll(ids []string) ([]timedTxRequest, error) {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-
-	requests := make([]timedTxRequest, 0, len(m.requests))
-	for _, request := range m.requests {
-		if len(ids) > 0 {
-			for _, id := range ids {
-				if request.Id == id {
-					requests = append(requests, *request)
-					break
-				}
-			}
-			continue
-		}
-		requests = append(requests, *request)
-	}
-	return requests, nil
-}
-
-func (m *txRequestsQueue) view(id string) (*domain.TxRequest, bool) {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-
-	request, ok := m.requests[id]
-	if !ok {
-		return nil, false
-	}
-
-	return &domain.TxRequest{
-		Id:        request.Id,
-		Inputs:    request.Inputs,
-		Receivers: request.Receivers,
-	}, true
-}
-
-func (m *txRequestsQueue) includesAny(outpoints []domain.VtxoKey) (bool, string) {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-
-	for _, out := range outpoints {
-		if _, exists := m.vtxos[out.String()]; exists {
-			return true, out.String()
-		}
-	}
-
-	return false, ""
-}
-
-type forfeitTxsMap struct {
-	lock    *sync.RWMutex
-	builder ports.TxBuilder
-
-	forfeitTxs      map[domain.VtxoKey]string
-	connectors      tree.TxTree
-	connectorsIndex map[string]domain.Outpoint
-	vtxos           []domain.Vtxo
-}
-
-func newForfeitTxsMap(txBuilder ports.TxBuilder) *forfeitTxsMap {
-	return &forfeitTxsMap{
-		lock:            &sync.RWMutex{},
-		builder:         txBuilder,
-		forfeitTxs:      make(map[domain.VtxoKey]string),
-		connectors:      nil,
-		connectorsIndex: nil,
-		vtxos:           nil,
-	}
-}
-
-func (m *forfeitTxsMap) init(connectors tree.TxTree, requests []domain.TxRequest) error {
-	vtxosToSign := make([]domain.Vtxo, 0)
-	for _, request := range requests {
-		for _, vtxo := range request.Inputs {
-			// If the vtxo is swept or is a note, it doens't require to be forfeited so we skip it
-			if !vtxo.RequiresForfeit() {
-				continue
-			}
-			vtxosToSign = append(vtxosToSign, vtxo)
-		}
-	}
-
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	m.vtxos = vtxosToSign
-	m.connectors = connectors
-
-	// init the forfeit txs map
-	for _, vtxo := range vtxosToSign {
-		m.forfeitTxs[vtxo.VtxoKey] = ""
-	}
-
-	// create the connectors index
-	connectorsIndex := make(map[string]domain.Outpoint)
-
-	if len(vtxosToSign) > 0 {
-		connectorsOutpoints := make([]domain.Outpoint, 0)
-
-		leaves := connectors.Leaves()
-		if len(leaves) == 0 {
-			return fmt.Errorf("no connectors found")
-		}
-
-		for _, n := range leaves {
-			connectorsOutpoints = append(connectorsOutpoints, domain.Outpoint{
-				Txid: n.Txid,
-				VOut: 0,
-			})
-		}
-
-		// sort lexicographically
-		sort.Slice(vtxosToSign, func(i, j int) bool {
-			return vtxosToSign[i].String() < vtxosToSign[j].String()
-		})
-
-		if len(vtxosToSign) > len(connectorsOutpoints) {
-			return fmt.Errorf("more vtxos to sign than outpoints, %d > %d", len(vtxosToSign), len(connectorsOutpoints))
-		}
-
-		for i, vtxo := range vtxosToSign {
-			connectorsIndex[vtxo.String()] = connectorsOutpoints[i]
-		}
-	}
-
-	m.connectorsIndex = connectorsIndex
-
-	return nil
-}
-
-func (m *forfeitTxsMap) sign(txs []string) error {
-	if len(txs) == 0 {
-		return nil
-	}
-
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	if len(m.vtxos) == 0 || len(m.connectors) == 0 {
-		return fmt.Errorf("forfeit txs map not initialized")
-	}
-
-	// verify the txs are valid
-	validTxs, err := m.builder.VerifyForfeitTxs(m.vtxos, m.connectors, txs, m.connectorsIndex)
-	if err != nil {
-		return err
-	}
-
-	for vtxoKey, txs := range validTxs {
-		if _, ok := m.forfeitTxs[vtxoKey]; !ok {
-			return fmt.Errorf("unexpected forfeit tx, vtxo %s is not in the batch", vtxoKey)
-		}
-		m.forfeitTxs[vtxoKey] = txs
-	}
-
-	return nil
-}
-
-func (m *forfeitTxsMap) reset() {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	m.forfeitTxs = make(map[domain.VtxoKey]string)
-	m.connectors = nil
-	m.connectorsIndex = nil
-	m.vtxos = nil
-}
-
-func (m *forfeitTxsMap) pop() ([]string, error) {
-	m.lock.Lock()
-	defer func() {
-		m.lock.Unlock()
-		m.reset()
-	}()
-
-	txs := make([]string, 0)
-	for vtxo, forfeit := range m.forfeitTxs {
-		if len(forfeit) == 0 {
-			return nil, fmt.Errorf("missing forfeit tx for vtxo %s", vtxo)
-		}
-		txs = append(txs, forfeit)
-	}
-
-	return txs, nil
-}
-
-func (m *forfeitTxsMap) allSigned() bool {
-	for _, txs := range m.forfeitTxs {
-		if len(txs) == 0 {
-			return false
-		}
-	}
-
-	return true
-}
-
-type offchainTxsMap struct {
-	lock        *sync.RWMutex
-	offchainTxs map[string]domain.OffchainTx
-	inputs      map[string]struct{}
-}
-
-func newOffchainTxsMap() *offchainTxsMap {
-	return &offchainTxsMap{
-		lock:        &sync.RWMutex{},
-		offchainTxs: make(map[string]domain.OffchainTx),
-		inputs:      make(map[string]struct{}),
-	}
-}
-
-func (m *offchainTxsMap) add(offchainTx domain.OffchainTx) {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-	m.offchainTxs[offchainTx.VirtualTxid] = offchainTx
-	for _, tx := range offchainTx.CheckpointTxs {
-		ptx, _ := psbt.NewFromRawBytes(strings.NewReader(tx), true)
-		for _, in := range ptx.UnsignedTx.TxIn {
-			m.inputs[in.PreviousOutPoint.String()] = struct{}{}
-		}
-	}
-}
-
-func (m *offchainTxsMap) remove(virtualTxid string) {
-	m.lock.Lock()
-	defer m.lock.Unlock()
-
-	offchainTx, ok := m.offchainTxs[virtualTxid]
-	if !ok {
-		return
-	}
-
-	for _, tx := range offchainTx.CheckpointTxs {
-		ptx, _ := psbt.NewFromRawBytes(strings.NewReader(tx), true)
-		for _, in := range ptx.UnsignedTx.TxIn {
-			delete(m.inputs, in.PreviousOutPoint.String())
-		}
-	}
-	delete(m.offchainTxs, virtualTxid)
-}
-
-func (m *offchainTxsMap) get(virtualTxid string) (domain.OffchainTx, bool) {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-	offchainTx, ok := m.offchainTxs[virtualTxid]
-	return offchainTx, ok
-}
-
-func (m *offchainTxsMap) includes(outpoint domain.VtxoKey) bool {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-	_, exists := m.inputs[outpoint.String()]
-	return exists
-}
 
 // onchainOutputs iterates over all the nodes' outputs in the vtxo tree and checks their onchain state
 // returns the sweepable outputs as ports.SweepInput mapped by their expiration time
